+++ config/module.config.php
@@ -4,6 +4,7 @@
  * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
+use ZF\HttpCache\DefaultETagGenerator;
 use ZF\HttpCache\HttpCacheListener;
 use ZF\HttpCache\HttpCacheListenerFactory;
 
@@ -89,6 +90,12 @@ return [
     //                'vary' => [
     //                    'override' => true, // Whether to override vary header if already present in response
     //                    'value'    => 'accept-encoding, x-requested-with',
+    //                ],
+    //
+    //                // Vary: http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.19
+    //                'etag' => [
+    //                    'override' => false, // Whether to override etag header if already present in response
+    //                    'generator' => '\ZF\HttpCache\DefaultETagGenerator', // what etag generator to use, must implement \ZF\HttpCache\EtagGeneratorInterface
     //                ],
     //            ],
     //        ],